train: move route index to TrainState, making TrainPath final
